backend/services/graphql-gateway/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
gunicorn>=20.1.0
strawberry-graphql>=0.171.1
httpx>=0.24.0
python-dotenv>=1.0.0

# MongoDB
motor>=3.1.1
pymongo>=4.3.3

# Performance and monitoring
prometheus-client>=0.16.0

# Utilities
pydantic>=1.10.7
pydantic-settings>=2.0.3
email-validator>=2.0.0

# Production
uvloop>=0.17.0; sys_platform != "win32"
httptools>=0.5.0
